+++ src/client/theme-default/index.ts
@@ -8,6 +8,7 @@ import { Theme } from 'vitepress'
 import Layout from './Layout.vue'
 import NotFound from './NotFound.vue'
 
+export { DefaultTheme } from './config'
 const theme: Theme = {
   Layout,
   NotFound
